Fix inject crate compile errors for Windows release build
Cast thread hide info length to u32, map pipe read errors to String, and
call export_rva_by_hash inside an unsafe block.

+++ inject/src/anti.rs
@@ -21,7 +21,7 @@ pub fn apply_stealth() {
             CURRENT_THREAD,
             THREAD_HIDE_FROM_DEBUGGER,
             &mut hide as *mut u8 as *mut std::ffi::c_void,
-            mem::size_of::<u8>(),
+            mem::size_of::<u8>() as u32,
         );
     }
 }
